moved automod from serverconfig.py to automod.py
